DOCS - Move util/settings files and add comments to game model
