+++ config/gameSettings.js
